pyproject.toml
[tool.poetry]
name = "imagecolor"
version = "2.0.0rc1"
description = "Image color extraction"
authors = ["Rhys Hansen <rhyshonline@gmail.com>"]
license = "MIT"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pylint = "^3.0.2"
Sphinx = "^7.2.6"
sphinx-rtd-theme = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
